# services/menu/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Menu Service API",
        "description": "Restaurant menu management service",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Menu Service"
        }
    ],
    "paths": {
...
        "/v1/menu/items/{id}/availability": {
            "post": {
                "summary": "Update item availability",
                "description": "Updates availability status of a menu item and reconciles with inventory",
                "operationId": "updateMenuItemAvailability",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Menu item ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "available"
                                ],
                                "properties": {
                                    "available": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Availability updated"
                    },
                    "404": {
                        "description": "Menu item not found"
                    }
                }
            }
        },
...
    },
...
}